Tighten the PHPStan setup beyond level max
level は「型推論をどこまで厳しく見るか」の軸でしかなく、ルールセットや
解析対象の範囲は別設定になっている。max のまま 0 件だったのは、見ていない
ところがあったため。

- phpstan-strict-rules / phpstan-deprecation-rules を追加する。後者は
  PHPStan 2.x のコアには入っていない。
- bin/dmm-api-client を paths に入れる。拡張子がないため paths に bin を
  足すだけでは 1 ファイルも解析されない。ファイル名で直接指定する。
- phpVersion を 8.3-8.5 の範囲で指定する。CI のテストは 8.3/8.4/8.5 を
  回しているのに、解析は lint ジョブの 8.3 相当しか見ていなかった。
- level: max はその時点の最高レベルへのエイリアスで、PHPStan の更新だけで
  意味が変わる。composer.lock を追跡していないぶん CI は毎回最新を入れる
  ので、10 と明示する。
- 追加してもほとんど指摘の出ないフラグをまとめて有効にする。

出た指摘への対応:

- $argv は register_argc_argv が無効だと未定義になる。probe.php と同じ
  注釈を bin にも置く。型は非空リストが実態に合う。
- 短縮三項は preg_split()/scandir()/glob() の false を潰す用途で使って
  いた。false 判定を明示に書き換える。
- ArticleTally の ID は数値文字列なので、PHP が配列キーを int に変換する。
  (string) のキャストは必要で、嘘だったのは array<string, int> のほう。
- Runner::BACKOFF は非空の定数配列で end() は必ず int を返すため、
  (int) のキャストは効いていない。
- $this->fail() は static メソッドの動的呼び出しになる。Assert::fail() は
  never を返すので、後続が到達不能であることも伝わる。

+++ tools/live-probe/src/Runner.php
@@ -321,7 +321,7 @@ final class Runner
     {
         $backoff = self::BACKOFF;
 
-        sleep($backoff[$attempt - 1] ?? (int) end($backoff));
+        sleep($backoff[$attempt - 1] ?? end($backoff));
     }
 
     private static function isRetryable(int $status): bool
